Add mechanism to ensure that the encrypted secnonce cannot be reused.
This is implemented using a 2-byte counter starting from 1.
This counter is increment during each GenerateSecnonce() execution.
A unique id based on this counter is saved in the encrypted secnonce that is exported, and also stored in the chip in a list of valid ids.

During Musig2SignHash(), the id recovered from the decrypted secnonce is checked against the list of valid ids.
If the id is found, the secnonce is considered valid and the id is removed from the list of valid ids to ensure it cannot be reused.
If the id is not found, the secnonce is considered invalid.

When the counter overflows, we invalidate all encrypted secnonces by generating a new MAC key, and we reset the list of valid ids.

The maximum size of the id list (BIP327_MAX_NB_ID) determines the maximum number of MuSig2 sessions that can be performed in parallel.
Newly generated ids overwrite old entries in the list when the list reaches its maximum size.

Compared to previous commit, we use a 2-byte counter which reduces the memory footprint.
After 65535 counter MuSig2 secnonce generations, the counter overflows and all existing secnonces are invalidated.
The counter then resets to 1 and Musig2GenerateNonce() can be invoked again.